requirements.txt
certifi==2025.1.31
charset-normalizer==3.4.1
dnspython==2.7.0
idna==3.10
requests==2.32.3
urllib3==2.3.0
